Rename offline_cache_manager.* to cache_manager.*
