+++ taichi/runtime/gfx/CMakeLists.txt
@@ -7,7 +7,7 @@ target_sources(gfx_runtime
     snode_tree_manager.cpp
     aot_module_builder_impl.cpp
     aot_module_loader_impl.cpp
-    offline_cache_manager.cpp
+    cache_manager.cpp
   )
 #TODO 4832, some dependencies here should not be required as they
 # are build requirements of other targets.
